requirements.txt
APScheduler==3.11.2
asgiref==3.11.1
Django==5.2.12
django-apscheduler==0.7.0
django-cors-headers==4.9.0
django-fsm==3.0.1
djangorestframework==3.16.1
dj-database-url==2.3.0
gunicorn==23.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.3
tzlocal==5.3.1
whitenoise==6.9.0
